reorganizing core classes and updating everything
