+++ pyrlap/pyrlap2/algorithms/__init__.py
@@ -1,2 +1,4 @@
 from pyrlap.pyrlap2.algorithms.vectorizedvalueiteration import \
     VectorizedValueIteration
+from pyrlap.pyrlap2.algorithms.laostar import LAOStar
+from pyrlap.pyrlap2.algorithms.lrtdp import LRTDP
